+++ src/Git_Updater/Init.php
@@ -50,8 +50,6 @@ class Init {
 	public function run() {
 		if ( ! static::is_heartbeat() ) {
 			$this->load_hooks();
-
-			new Ignore( 'git-updater-pro', 'git-updater-pro/git-updater-pro.php' );
 		}
 
 		if ( static::is_wp_cli() ) {