+++ CHANGES.md
@@ -2,7 +2,7 @@
 * re-integrate Git Updater PRO
 * integrate Git Updater Additions
 * add Authentication tab
-* add option to dynamically load API tabs, previous default
+* load API tabs of installed/active API plugins
 
 #### 11.1.4 / 2022-10-04 (hotfix)
 * initialize Appsero SDK from `plugins_loaded` hook